moved all images to assets
