+++ src/Components/Home/homepage.css
@@ -1,5 +1,5 @@
 .banner {
-  background-image: url("/src/Components/Home/Images/banner.png");
+  background-image: url("/src/assets/Images/banner.png");
   width: 100%;
   height: 18rem;
   object-fit: cover;
