+++ src/backend/db/categories.js
@@ -1,8 +1,8 @@
 import { v4 as uuid } from "uuid";
-import makeup from "../../Components/Home/Images/makeup.jpg";
-import skincare from "../../Components/Home/Images/skin.jpg";
-import haircare from "../../Components/Home/Images/hair.jpg";
-import fragrances from "../../Components/Home/Images/Fragrances.jpg";
+import makeup from "../../assets/Images/makeup.jpg";
+import skincare from "../../assets/Images/skin.jpg";
+import haircare from "../../assets/Images/hair.jpg";
+import fragrances from "../../assets/Images/Fragrances.jpg";
 /**
  * Category Database can be added here.
  * You can add category of your wish with different attributes
