+++ src/backend/db/topbrands.js
@@ -1,7 +1,7 @@
-import brand1 from "../../Components/Home/Images/brand1.jpg";
-import brand2 from "../../Components/Home/Images/brand2.jpg";
-import brand3 from "../../Components/Home/Images/brand3.jpg";
-import brand4 from "../../Components/Home/Images/brand4.jpg";
+import brand1 from "../../assets/Images/brand1.jpg";
+import brand2 from "../../assets/Images/brand2.jpg";
+import brand3 from "../../assets/Images/brand3.jpg";
+import brand4 from "../../assets/Images/brand4.jpg";
 
 
 const topbrandslist = [brand1, brand2, brand3, brand4];
